+++ src/main/java/colorful/starbucks/auth/application/AuthService.java
@@ -8,23 +8,11 @@ public interface AuthService {
 
     void signUp(MemberSignUpRequestDto memberSignUpRequestDto);
 
-    boolean isEmailDuplicated(String email);
-
     UserDetails loadUserByUuid(String uuid);
 
     MemberSignInResponseDto signIn(MemberSignInRequestDto memberSignInRequestDto);
 
     AccessTokenResponseDto reIssueAccessToken(RefreshTokenRequestDto refreshTokenRequestDto);
-
-    MemberEmailFindResponseDto findEmail(MemberEmailFindRequestDto memberEmailFindRequestDto);
-
-    MemberPasswordResetResponseDto findPassword(MemberPasswordResetRequestDto memberPasswordResetRequestDto);
-
-    MemberSignInResponseDto kakaoSignIn(KakaoSignInRequestDto kakaoSignInRequestDto);
-
-    EmailCodeSendResponseDto sendEmail(EmailCodeSendRequestDto emailCodeSendRequestDto);
-
-    void verifyEmailCode(EmailVerifyCodeRequestDto emailVerifyCodeRequestDto);
 
     void signOut(MemberSignOutRequestDto memberSignOutRequestDto);
