+++ src/main/java/colorful/starbucks/auth/application/AuthServiceImpl.java
@@ -1,17 +1,18 @@
 package colorful.starbucks.auth.application;
 
 import colorful.starbucks.auth.domain.CustomUserDetails;
-import colorful.starbucks.member.domain.Member;
-import colorful.starbucks.member.domain.MemberLevel;
-import colorful.starbucks.auth.domain.SignType;
-import colorful.starbucks.auth.dto.request.*;
-import colorful.starbucks.auth.dto.response.*;
+import colorful.starbucks.auth.dto.request.MemberSignInRequestDto;
+import colorful.starbucks.auth.dto.request.MemberSignOutRequestDto;
+import colorful.starbucks.auth.dto.request.MemberSignUpRequestDto;
+import colorful.starbucks.auth.dto.request.RefreshTokenRequestDto;
+import colorful.starbucks.auth.dto.response.AccessTokenResponseDto;
+import colorful.starbucks.auth.dto.response.MemberSignInResponseDto;
 import colorful.starbucks.auth.infrastructure.AuthRepository;
 import colorful.starbucks.common.exception.BaseException;
 import colorful.starbucks.common.jwt.JwtTokenProvider;
 import colorful.starbucks.common.response.ResponseStatus;
-import colorful.starbucks.common.service.EmailService;
-import colorful.starbucks.common.util.UuidGenerator;
+import colorful.starbucks.common.util.TokenGenerator;
+import colorful.starbucks.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -32,11 +33,9 @@ public class AuthServiceImpl implements AuthService {
     private final PasswordEncoder passwordEncoder;
     private final JwtTokenProvider jwtTokenProvider;
     private final AuthenticationManager authenticationManager;
-    private final EmailService emailService;
-    private final KakaoApiService kakaoApiService;
     private final UserDetailsService userDetailsService;
-    private final EmailAuthRedisService emailAuthRedisService;
     private final RefreshTokenRedisService refreshTokenRedisService;
+    private final TokenGenerator tokenGenerator;
 
 
     @Override
@@ -53,41 +52,9 @@ public class AuthServiceImpl implements AuthService {
         Authentication authentication = authenticationManager.authenticate(
                 toAuthenticationToken(memberSignInRequestDto)
         );
-        return generateTokensAndSaveRefresh(authentication);
+        return tokenGenerator.issueTokens(authentication);
     }
 
-    @Transactional
-    @Override
-    public MemberSignInResponseDto kakaoSignIn(KakaoSignInRequestDto kakaoSignInRequestDto) {
-        KakaoUserInfo userInfo = kakaoSignInRequestDto.fetchUserInfo(kakaoApiService);
-
-        if (userInfo.getEmail() == null || userInfo.getEmail().isBlank()) {
-            throw new BaseException(ResponseStatus.INVALID_EMAIL_ADDRESS);
-        }
-
-        Member member = findOrCreateKakaoMember(userInfo);
-        UserDetails userDetails = new CustomUserDetails(member);
-
-        Authentication authentication = new UsernamePasswordAuthenticationToken(
-                userDetails, null, userDetails.getAuthorities()
-        );
-
-        return generateTokensAndSaveRefresh(authentication);
-    }
-
-    private MemberSignInResponseDto generateTokensAndSaveRefresh(Authentication authentication) {
-        String accessToken = createAccessToken(authentication);
-        String refreshToken = createRefreshToken(authentication);
-
-        String uuid = ((CustomUserDetails) authentication.getPrincipal()).getMemberUuid();
-        refreshTokenRedisService.saveRefreshToken(
-                uuid,
-                refreshToken,
-                jwtTokenProvider.getRefreshTokenExpireTime()
-        );
-
-        return MemberSignInResponseDto.from(accessToken, refreshToken);
-    }
 
     private UsernamePasswordAuthenticationToken toAuthenticationToken(MemberSignInRequestDto memberSignInRequestDto) {
         return new UsernamePasswordAuthenticationToken(memberSignInRequestDto.getEmail(), memberSignInRequestDto.getPassword());
@@ -98,14 +65,11 @@ public class AuthServiceImpl implements AuthService {
     @Override
     public void signUp(MemberSignUpRequestDto memberSignUpRequestDto) {
         authRepository.findByEmail(memberSignUpRequestDto.getEmail()).ifPresent(
-                (member) -> { throw new BaseException(ResponseStatus.DUPLICATED_USER); }
+                (member) -> {
+                    throw new BaseException(ResponseStatus.DUPLICATED_USER);
+                }
         );
         authRepository.save(memberSignUpRequestDto.toEntityWithEncodePassword(passwordEncoder));
-    }
-
-    @Override
-    public boolean isEmailDuplicated(String email) {
-        return authRepository.existsByEmail(email);
     }
 
     @Transactional
@@ -122,71 +86,6 @@ public class AuthServiceImpl implements AuthService {
         return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
     }
 
-    private String createAccessToken(Authentication authentication) {
-        return jwtTokenProvider.generateAccessToken(authentication);
-    }
-
-    private String createRefreshToken(Authentication authentication) {
-        return jwtTokenProvider.generateRefreshToken(authentication);
-    }
-
-    @Transactional
-    @Override
-    public EmailCodeSendResponseDto sendEmail(EmailCodeSendRequestDto emailCodeSendRequestDto) {
-        String code = emailCodeSendRequestDto.codeGenerator();
-        emailAuthRedisService.saveCode(emailCodeSendRequestDto.getEmail(), code);
-        emailService.sendEmailCode(emailCodeSendRequestDto.getEmail(), code);
-        return EmailCodeSendResponseDto.from(code);
-    }
-
-    @Transactional
-    @Override
-    public void verifyEmailCode(EmailVerifyCodeRequestDto emailVerifyCodeRequestDto) {
-        if (emailAuthRedisService.verifyCode(emailVerifyCodeRequestDto.getEmail(), emailVerifyCodeRequestDto.getCode())) {
-            emailAuthRedisService.deleteCode(emailVerifyCodeRequestDto.getEmail());
-        } else {
-            throw new BaseException(ResponseStatus.INVALID_AUTH_CODE);
-        }
-    }
-
-    @Transactional
-    @Override
-    public MemberPasswordResetResponseDto findPassword(MemberPasswordResetRequestDto memberPasswordResetRequestDto) {
-        if (!authRepository.existsByEmail(memberPasswordResetRequestDto.getEmail())) {
-            throw new BaseException(ResponseStatus.NO_EXIST_USER);
-        }
-
-        Member member = memberPasswordResetRequestDto.findMatchingMember(authRepository)
-                .orElseThrow(() -> new BaseException(ResponseStatus.NO_EXIST_USER));
-
-        memberPasswordResetRequestDto.generateTempPassword(passwordEncoder);
-        member.updatePassword(memberPasswordResetRequestDto.getEncodedPassword());
-        emailService.sendTempPassword(member.getEmail(), memberPasswordResetRequestDto.getTempPassword());
-
-        return MemberPasswordResetResponseDto.from("임시 비밀번호가 이메일로 전송 되었습니다.");
-    }
-
-
-    private Member findOrCreateKakaoMember(KakaoUserInfo kakaoUserInfo) {
-        return authRepository.findBySignTypeAndSocialId(SignType.KAKAO, kakaoUserInfo.getId())
-                .orElseGet(() -> authRepository.save(
-                        Member.builder()
-                                .signType(SignType.KAKAO)
-                                .socialId(kakaoUserInfo.getId())
-                                .email(kakaoUserInfo.getEmail())
-                                .memberUuid(UuidGenerator.generateUuid())
-                                .memberLevel(MemberLevel.WHITE)
-                                .build()
-                ));
-    }
-
-    @Transactional
-    @Override
-    public MemberEmailFindResponseDto findEmail(MemberEmailFindRequestDto memberEmailFindRequestDto) {
-        Member member = memberEmailFindRequestDto.findByMemberNameAndPhoneNumber(authRepository)
-                .orElseThrow(() -> new BaseException(ResponseStatus.NO_EXIST_USER));
-        return MemberEmailFindResponseDto.from(member.getEmail());
-    }
 
     @Transactional
     @Override
