+++ src/main/java/colorful/starbucks/auth/dto/request/MemberEmailFindRequestDto.java
@@ -1,7 +1,6 @@
 package colorful.starbucks.auth.dto.request;
 
 import colorful.starbucks.member.domain.Member;
-import colorful.starbucks.auth.infrastructure.AuthRepository;
 import colorful.starbucks.auth.vo.request.MemberEmailFindRequestVo;
 import lombok.Builder;
 import lombok.Getter;
@@ -27,11 +26,5 @@ public class MemberEmailFindRequestDto {
                 .phoneNumber(memberEmailFindRequestVo.getPhoneNumber())
                 .build();
     }
-
-    public Optional<Member>  findByMemberNameAndPhoneNumber(AuthRepository authRepository) {
-        return authRepository.findByMemberNameAndPhoneNumber(memberName, phoneNumber);
-
-    }
-
 
 }
