+++ src/main/java/colorful/starbucks/auth/infrastructure/AuthRepository.java
@@ -8,17 +8,10 @@ import java.util.Optional;
 
 public interface AuthRepository extends JpaRepository<Member, Long> {
 
-    boolean existsByEmail(String email);
-
     Optional<Member> findByMemberUuid(String uuid);
 
     Optional<Member> findByEmail(String email);
 
-    Optional<Member> findByMemberNameAndPhoneNumber(String memberName, String phoneNumber);
-
-    Optional<Member> findByEmailAndMemberNameAndPhoneNumber(String email, String memberName, String phoneNumber);
-
     Optional<Member> findBySignTypeAndSocialId(SignType signInType, String socialId);
-
 
 }
