+++ src/main/java/colorful/starbucks/auth/presentation/AuthController.java
@@ -21,12 +21,7 @@ public class AuthController {
         return ApiResponse.ok("회원가입이 완료되었습니다.", null);
     }
 
-    @GetMapping("/email/exists")
-    public ApiResponse<Boolean> checkEmailDuplication(@RequestParam String email) {
-        return ApiResponse.ok(
-                "이메일 중복 체크를 완료하였습니다.",
-                authService.isEmailDuplicated(email));
-    }
+
 
     @PostMapping("/sign-in")
     public ApiResponse<MemberSignInResponseVo> signIn(@RequestBody MemberSignInRequestVo memberSignInRequestVo) {
@@ -42,55 +37,16 @@ public class AuthController {
                 authService.reIssueAccessToken(RefreshTokenRequestDto.from(refreshTokenRequestVo)).toVo());
     }
 
-    @PostMapping("/email")
-    public ApiResponse<MemberEmailFindResponseVo> findEmail(@RequestBody MemberEmailFindRequestVo memberEmailFindRequestVo) {
-        return ApiResponse.ok(
-                "이메일 찾기를 완료하였습니다.",
-                authService.findEmail(MemberEmailFindRequestDto.from(memberEmailFindRequestVo)).toVo());
-    }
-
-    @PostMapping("/password")
-    public ApiResponse<MemberPasswordResetResponseVo> resetPassword(@RequestBody MemberPasswordResetRequestVo memberPasswordResetRequestVo) {
-        return ApiResponse.ok(
-                "비밀번호 찾기를 완료하였습니다.",
-                authService.findPassword(MemberPasswordResetRequestDto.from(memberPasswordResetRequestVo)).toVo());
-    }
-
     @DeleteMapping("/sign-out")
     public ApiResponse<Void> signOut(@RequestBody MemberSignOutRequestDto dto) {
         authService.signOut(dto);
         return ApiResponse.ok("로그아웃을 완료했습니다",null);
     }
 
-    @PostMapping("/kakao")
-    public ApiResponse<MemberSignInResponseVo> kakaoSignIn(@RequestBody KakaoSignInRequestVo kakaoSignInRequestVo) {
-        return ApiResponse.ok(
-                "로그인에 성공하였습니다.",
-                authService.kakaoSignIn(KakaoSignInRequestDto.from(kakaoSignInRequestVo)).toVo());
-    }
-
-    @PostMapping("/email/send-code")
-    public ApiResponse<EmailCodeSendResponseVo> sendEmail(@RequestBody EmailCodeSendRequestVo emailCodeSendRequestVo) {
-        return ApiResponse.ok(
-                "이메일 인증 번호를 전송하였습니다",
-                authService.sendEmail(EmailCodeSendRequestDto.from(emailCodeSendRequestVo)).toVo()
-        );
-    }
-
-    @PostMapping("/email/verify-code")
-    public ApiResponse<String> verifyEmailCode(@RequestBody EmailVerifyCodeRequestVo emailVerifyCodeRequestVo) {
-        authService.verifyEmailCode(EmailVerifyCodeRequestDto.from(emailVerifyCodeRequestVo));
-        return ApiResponse.ok("이메일 인증이 완료되었습니다.");
-    }
 
 
 
 
-    //TODO 백엔드 카카오 로그인 인가 코드 api
-    @GetMapping("/kakao")
-    public String kakaoRedirect(@RequestParam String code) {
 
-        System.out.println("카카오 인가코드 = " + code);
-        return "OK";
-    }
+
 }
