+++ src/main/java/colorful/starbucks/member/application/TermsService.java
@@ -1,0 +1,14 @@
+package colorful.starbucks.member.application;
+
+import colorful.starbucks.member.dto.request.TermsCreateRequestDto;
+import colorful.starbucks.member.dto.response.TermsResponseDto;
+
+import java.util.List;
+
+public interface TermsService {
+
+
+     void createTerms(TermsCreateRequestDto termsCreateRequestDto);
+
+     List<TermsResponseDto> getTerms();
+}
