+++ src/main/java/colorful/starbucks/member/domain/TermsAgreement.java
@@ -1,4 +1,4 @@
-package colorful.starbucks.auth.domain;
+package colorful.starbucks.member.domain;
 
 import colorful.starbucks.common.entity.BaseEntity;
 import jakarta.persistence.*;
