+++ src/main/java/colorful/starbucks/member/dto/response/TermsAgreementResponseDto.java
@@ -1,4 +1,4 @@
-package colorful.starbucks.auth.dto.response;
+package colorful.starbucks.member.dto.response;
 
 import lombok.Getter;
