+++ src/main/java/colorful/starbucks/member/infrastructure/MemberRepository.java
@@ -9,4 +9,10 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findAllByMemberUuid(String memberUuid);
+
+    Optional<Member> findByEmailAndMemberNameAndPhoneNumber(String email, String memberName, String phoneNumber);
+
+    Optional<Member> findByMemberNameAndPhoneNumber(String memberName, String phoneNumber);
+
+    boolean existsByEmail(String email);
 }
