+++ src/main/java/colorful/starbucks/member/infrastructure/TermsAgreementRepository.java
@@ -1,6 +1,6 @@
-package colorful.starbucks.auth.infrastructure;
+package colorful.starbucks.member.infrastructure;
 
-import colorful.starbucks.auth.domain.TermsAgreement;
+import colorful.starbucks.member.domain.TermsAgreement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TermsAgreementRepository extends JpaRepository<TermsAgreement, Long> {
