+++ src/main/java/colorful/starbucks/member/infrastructure/TermsRepository.java
@@ -1,6 +1,6 @@
-package colorful.starbucks.auth.infrastructure;
+package colorful.starbucks.member.infrastructure;
 
-import colorful.starbucks.auth.domain.Terms;
+import colorful.starbucks.member.domain.Terms;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TermsRepository extends JpaRepository<Terms, Long> {
